Tighten up ComponentModule type
